requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
db-dtypes==1.4.2
flasgger==0.9.7.1
Flask==3.1.0
flask-sock==0.7.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
MarkupSafe==3.0.2
mistune==3.1.2
narwhals==1.30.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
plotly==6.0.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vanna==0.7.6
Werkzeug==3.1.3
wsproto==1.2.0
